requirements.txt
# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Vector Store
chromadb>=0.4.22

# Agent Framework
strands-agents>=0.1.0
strands-agents-tools>=0.1.0

# LLM Providers
anthropic>=0.18.0
openai>=1.10.0
litellm>=1.20.0

# Embeddings
sentence-transformers>=2.3.0

# Document Processing
pypdf>=3.17.0
python-docx>=1.1.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
markdown>=3.5.0

# Hybrid Search
rank-bm25>=0.2.2

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.6
aiofiles>=23.2.1

# SSE Support
sse-starlette>=1.8.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
httpx>=0.26.0
